requirements.txt
torch>=2.2
torchvision
pandas
scikit-learn
matplotlib
albumentations
streamlit
pillow
pytest
